readInsertData/requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
async-timeout==5.0.1
colorzero==2.0
dotenv==0.9.9
exceptiongroup==1.2.2
gpiozero==2.0.1
hiredis==3.1.0
idna==3.10
lgpio==0.2.2.0
multidict==6.1.0
pamqp==3.3.0
pika==1.3.2
propcache==0.3.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
redis==5.2.1
rpi-lgpio==0.6
RPi.GPIO==0.7.1
RPLCD==1.3.1
smbus2==0.5.0
yarl==1.18.3
